+++ AD-P03_SpringMongoDB/src/main/kotlin/es/dam/adp03_springmongodb/models/Maquina.kt
@@ -19,7 +19,7 @@ data class Maquina(
     val modelo: String,
     @Contextual
     val fechaAdquisicion: LocalDate,
-    val numeroSerie: Int,
+    val numeroSerie: Long,
     val tipo: TipoMaquina,
     val descripcion: String
 )
